Moves ParseResultExtensions into main project
